misc shortdomain deletion updates etc

+++ app/Actions/Domains/DeleteShortDomain.php
@@ -17,22 +17,33 @@ class DeleteShortDomain
 
     /** 
      * works, but..
-     *  @todo refactor with jobs in a queue? error handling
+     *  @todo refactor with jobs in a queue?
+     * clear any existing createHostname etc related jobs in queue?
      */
     public function delete(Domain $domain)
     {
 
-        $keys = [];
-        foreach($domain->links as $link)
-        {
-            $keys[] = $link->auto_url;
-        }
+        if(!$domain->shortdomain)
+            return false;
+
 
         $this->cloudflare->deleteCustomHostname( $domain->shortdomain->cf_host_identifier );
-        $this->cloudflare->deleteWorkerRoute( $domain->shortdomain->cf_route_identifier );
-        
-        if(!empty($keys))
-            $this->cloudflare->bulkDeleteShortlinks( $keys );
+
+        if($domain->shortdomain->status === 'active') {
+
+            $keys = [];
+            foreach($domain->links as $link)
+            {
+                $keys[] = $link->auto_url;
+            }
+
+            if($domain->shordomain->cf_route_identifier)
+                $this->cloudflare->deleteWorkerRoute( $domain->shortdomain->cf_route_identifier );
+            
+            if(!empty($keys))
+                $this->cloudflare->bulkDeleteShortlinks( $keys );
+
+        }
 
         $domain->shortdomain->delete();
